Added Jest and Sinon as for automated testing
I added Jest because it is addressed in the issue #12. It is also
the growing standard, and easy for individuals to pick up if they
have never touched unit testing before. Sinon is a non-framework
dependant JavaScript mocking library. This means either Jest or
Sinon could be swapped out more easilly later.

I have set the current coverage cut off to a typically low 30%.
This is to make unit tests not absolutely required for participation
in the project.

I've taken the libery of writing the first function's tests to
help test the test suite, and provide a template for people to
checkout if they want to write laters. The function is pipe().

I've written robust documentation and put it in README.md